fix: run backstop test even if reference fails
I'm seeing a 'backstop reference' end in an error but it's still
displaying the message to run 'backstop test'. I'm not sure what the
error is (because the error reporting converts the error object to a
string so we just see '[object Object]'!) but nothing looks obviously
awful about the reference run, so change behavior to run 'backstop test'
even if 'backstop reference' exits with a failure code.